add ability for tags to hold hotkey info
This couples user interface info to the data-centric TagMap container. In my opinion, this is acceptable to avoid extra objects that link hotkeys to tags that in turn need their own serialization and deserialization and error-checking and such. The only benefit of doing so would be to position the TagMap interface as an independent library, which is not currently in the cards.