swap left and right panels in main frame(!)
I'll try this out for a bit. I think it flows better by conveying that the file comes "first" and the tags require the file to be present, so they come "second," matching left-to-right reading order. (Not everyone reads left to right, of course. It would be interesting to get input from those familiar with right-to-left scripts.)