move directory navigation above media
This is an experiment. It's not bad... I like how it emphasizes the Load File button a bit more and brings it adjacent to the media display. A downside of this arrangement is that the sightlines between media and tags are now diagonal. There's a certain ergonomic convenience to having the media next to the list of tags so one can glance between the two to determine whether tags are a good fit for what is being shown.