hack to preserve desired auto-fit for all columns
Without this maneuver, the last column gets stretched to fill the remaining control width whenever the control is redrawn. This is the documented behavior of `wxLIST_AUTOSIZE_HEADER`, though personally I think it's a poor design choice for a flag whose primary purpose is to set the column width to the width of the header text.