highlight summary file on main frame file select
Already it was the case that selecting a summary frame item would highlight the main frame item (if present), but now the opposite happens also. This commit also fixes some subtle bugs regarding procuring paths using list control item indices. The issue is that the indices would change after sort operations, so we couldn't rely on old info to do new actions. This could have been bad, especially in the context of removing (unintended) files from a project with a bulk action...